Introduce Repository concept along with specs and module refactoring
An initialize file is introduced to aid in loading up the required
classes and libraries.

A spec_helper and basic specs (for Repository) exists, along with a
RepositoryHelper to handle temporary git repos.

Refactor classes to use nested modules.

Change lib/git_evolution to be a module, thus executables had to change.