Fix git log for when no range was specified
When no range was specified the following arguments were passed to git
log: `-L,:<file>`. While it was still valid for the whole file it was
restricted to the bounds of the current file's state (i.e., 1:10, which
then restricted how far back the history will go by the nature of the
`-L` argument.

This commit conditionally removes the `-L` argument if there is no
starting or ending lines specified. This allows the history to go 'all'
the way back while following the file movements.

Slight refactor to improve readability of the ever growing git command.